+++ Loxodon.Framework/Assets/LoxodonFramework/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Requires Unity 2018.4 or higher.
 
+## [2.1.0] - 2021-07-22
+### Added feature that does not trigger interaction action if the view is disable.
+### Added window state broadcast feature to support subscribing to state change events for all windows via messenger component.
+### Supports using the "CanvasGroup.blocksRaycasts" instead of the "CanvasGroup.interactable" to control the interactivity of the view  
+
+
 ## [2.0.5] - 2020-09-07
 ### Optimize data binding on IOS platform, replace reflection function by creating delegate at runtime.